improvement(doc): restructuring template and shared functions folder
