+++ packages/doc/src/components/CodeBlock/shared/functions.js
@@ -1,0 +1,12 @@
+const injectImport = (imports, component) => {
+  const index = imports.indexOf(` } from '@gympass/yoga'`);
+  const importsWithComponent = [
+    imports.slice(0, index),
+    `, ${component}`,
+    imports.slice(index),
+  ].join('');
+
+  return importsWithComponent;
+};
+
+export default injectImport;
