+++ packages/doc/src/components/CodeBlock/shared/index.js
@@ -1,0 +1,3 @@
+import injectImport from './functions';
+
+export default injectImport;
