+++ packages/doc/src/components/CodeBlock/shared/templates/index.js
@@ -1,0 +1,4 @@
+import web from './web';
+import native from './native';
+
+export { web, native };
