+++ packages/doc/src/components/CodeBlock/shared/templates/web.js
@@ -1,0 +1,16 @@
+import injectImport from '..';
+
+const web = (imports, code) => `import React from 'react';
+import ReactDOM from 'react-dom';
+${injectImport(imports, 'ThemeProvider')}
+
+const App = () => <ThemeProvider>
+${code}
+</ThemeProvider>
+
+ReactDOM.render(
+  <App />,
+  document.getElementById('root')
+);`;
+
+export default web;
